Gomail OpenFile now support original file name
Original file name is used to detect the extension and fill up the
correct email header field. The first argument filename of OpenFile
is used to get the content of the file. If original file name is not
supply then gomail will assume both the path name and original file
name is same value.