Make "message/rfc822" Send As Regular Attachment File
This is make sure message/rfc822 is sending as regular attachment
content-type header "application/octet-stream" which is allowed
to use base64 encoding. Without base64 encoding, any file with
content-type message/rfc822 possible need to be checked and
modified to be able to send as 7bit or 8bit or binary. The
downside will be, this attachment content-type can not directly be
read-ed by any email client that capable reading message/rfc822
content-type. The good thing base64 can be used and no SMTP will
interfere/modify the message/rfc822 attachment.